Rename 'functions' package to 'forms'.
