--- README.adoc
@@ -1,1 +1,0 @@
-src/site/asciidoc/quickstart.adoc
