+++ pom.xml
@@ -1,13 +1,13 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.github.dakusui</groupId>
-    <artifactId>${project.name}</artifactId>
+    <artifactId>pcond</artifactId>
     <packaging>jar</packaging>
     <version>3.0.0-SNAPSHOT</version>
     <description>A Small Java Library for Design by Contract style programming with printable preconditions.
     </description>
 
-    <name>pcond</name>
+    <name>${project.artifactId}</name>
     <url>http://dakusui.github.io/${project.name}</url>
 
     <licenses>
@@ -257,6 +257,21 @@
                 <groupId>org.codehaus.mojo</groupId>
                 <artifactId>exec-maven-plugin</artifactId>
                 <executions>
+                    <execution><!-- Remove diagram caches -->
+                        <id>remove diag-*.png files</id>
+                        <phase>generate-sources</phase>
+                        <goals>
+                            <goal>exec</goal>
+                        </goals>
+                        <configuration>
+                            <executable>bash</executable>
+                            <arguments>
+                                <argument>
+                                    ${basedir}/src/main/build_tools/remove-adoc-diags.sh
+                                </argument>
+                            </arguments>
+                        </configuration>
+                    </execution>
                     <execution><!-- Run our version calculation script -->
                         <id>package-info.java generation</id>
                         <phase>generate-sources</phase>
