+++ src/main/build_tools/remove-adoc-diags.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+set -eu
+
+rm diag-*.png || echo "Let's go ahead." >&2
