+++ src/main/java/com/github/dakusui/pcond/forms/Printables.java
@@ -1,4 +1,4 @@
-package com.github.dakusui.pcond.functions;
+package com.github.dakusui.pcond.forms;
 
 
 import com.github.dakusui.pcond.core.printable.PrintableFunctionFactory;
