+++ src/site/asciidoc/attributes.adoc
@@ -1,2 +1,2 @@
-:pcond-version: 2.0.1
+:pcond-version: 3.0.0
