+++ src/test/java/com/github/dakusui/pcond/compilability/BooleanTest.java
@@ -3,7 +3,7 @@ package com.github.dakusui.pcond.compilability;
 import org.junit.Test;
 
 import static com.github.dakusui.pcond.Preconditions.requireArgument;
-import static com.github.dakusui.pcond.functions.Predicates.*;
+import static com.github.dakusui.pcond.forms.Predicates.*;
 
 public class BooleanTest {
   @Test(expected = IllegalArgumentException.class)
