+++ src/test/java/com/github/dakusui/pcond/examples/ExampleUT.java
@@ -8,7 +8,7 @@ import org.junit.Test;
 
 import static com.github.dakusui.pcond.TestAssertions.assertThat;
 import static com.github.dakusui.pcond.TestAssertions.assumeThat;
-import static com.github.dakusui.pcond.functions.Predicates.*;
+import static com.github.dakusui.pcond.forms.Predicates.*;
 import static com.github.dakusui.pcond.utils.TestMethodExpectation.Result.*;
 
 @TestClassExpectation({
