+++ src/test/java/com/github/dakusui/pcond/sandbox/Sandbox.java
@@ -4,7 +4,7 @@ import org.junit.Test;
 
 import java.util.function.Predicate;
 
-import static com.github.dakusui.pcond.functions.Predicates.*;
+import static com.github.dakusui.pcond.forms.Predicates.*;
 
 public class Sandbox {
   @Test
