+++ src/test/java/com/github/dakusui/pcond/ut/ExperimentalsTest.java
@@ -3,8 +3,8 @@ package com.github.dakusui.pcond.ut;
 import com.github.dakusui.pcond.core.context.Context;
 import com.github.dakusui.pcond.core.currying.CurriedFunction;
 import com.github.dakusui.pcond.core.printable.PrintableFunctionFactory;
-import com.github.dakusui.pcond.functions.Experimentals;
-import com.github.dakusui.pcond.functions.Functions;
+import com.github.dakusui.pcond.forms.Experimentals;
+import com.github.dakusui.pcond.forms.Functions;
 import com.github.dakusui.pcond.internals.InternalException;
 import com.github.dakusui.pcond.provider.PreconditionViolationException;
 import com.github.dakusui.pcond.utils.ut.TestBase;
@@ -17,9 +17,9 @@ import java.util.function.Function;
 import java.util.function.Predicate;
 
 import static com.github.dakusui.pcond.Preconditions.require;
-import static com.github.dakusui.pcond.functions.Experimentals.*;
-import static com.github.dakusui.pcond.functions.Functions.*;
-import static com.github.dakusui.pcond.functions.Predicates.*;
+import static com.github.dakusui.pcond.forms.Experimentals.*;
+import static com.github.dakusui.pcond.forms.Functions.*;
+import static com.github.dakusui.pcond.forms.Predicates.*;
 import static com.github.dakusui.pcond.internals.InternalUtils.wrapIfNecessary;
 import static com.github.dakusui.pcond.ut.ExperimentalsTest.Utils.areEqual;
 import static com.github.dakusui.pcond.ut.ExperimentalsTest.Utils.stringEndsWith;
