+++ src/test/java/com/github/dakusui/pcond/ut/ParameterizedFunctionsTest.java
@@ -1,6 +1,6 @@
 package com.github.dakusui.pcond.ut;
 
-import com.github.dakusui.pcond.functions.Functions;
+import com.github.dakusui.pcond.forms.Functions;
 import org.hamcrest.CoreMatchers;
 import org.hamcrest.Matcher;
 import org.junit.Test;
