+++ src/test/java/com/github/dakusui/pcond/ut/PostconditionsTest.java
@@ -2,7 +2,7 @@ package com.github.dakusui.pcond.ut;
 
 import com.github.dakusui.pcond.Postconditions;
 import com.github.dakusui.pcond.provider.AssertionProvider;
-import com.github.dakusui.pcond.functions.Predicates;
+import com.github.dakusui.pcond.forms.Predicates;
 import com.github.dakusui.pcond.provider.PostconditionViolationException;
 import com.github.dakusui.pcond.utils.ut.TestBase;
 import org.junit.Test;
