+++ src/test/java/com/github/dakusui/pcond/ut/PreconditionsTest.java
@@ -1,14 +1,14 @@
 package com.github.dakusui.pcond.ut;
 
 import com.github.dakusui.pcond.Preconditions;
-import com.github.dakusui.pcond.functions.Functions;
-import com.github.dakusui.pcond.functions.Predicates;
+import com.github.dakusui.pcond.forms.Functions;
+import com.github.dakusui.pcond.forms.Predicates;
 import com.github.dakusui.pcond.provider.PreconditionViolationException;
 import com.github.dakusui.pcond.utils.ut.TestBase;
 import org.hamcrest.CoreMatchers;
 import org.junit.Test;
 
-import static com.github.dakusui.pcond.functions.Predicates.isNotNull;
+import static com.github.dakusui.pcond.forms.Predicates.isNotNull;
 import static com.github.dakusui.pcond.utils.TestUtils.firstLineOf;
 import static org.hamcrest.CoreMatchers.*;
 import static org.junit.Assert.assertNotNull;
