+++ src/test/java/com/github/dakusui/pcond/ut/TransformingPredicateTest.java
@@ -2,8 +2,8 @@ package com.github.dakusui.pcond.ut;
 
 import com.github.dakusui.pcond.core.printable.PrintableFunctionFactory;
 import com.github.dakusui.pcond.core.printable.PrintablePredicateFactory;
-import com.github.dakusui.pcond.functions.Functions;
-import com.github.dakusui.pcond.functions.Predicates;
+import com.github.dakusui.pcond.forms.Functions;
+import com.github.dakusui.pcond.forms.Predicates;
 import com.github.dakusui.pcond.utils.ut.TestBase;
 import org.junit.Test;
