+++ src/test/java/com/github/dakusui/pcond/ut/bugfixes/Issue42Test.java
@@ -6,8 +6,8 @@ import java.util.*;
 import java.util.function.Function;
 
 import static com.github.dakusui.pcond.Preconditions.require;
-import static com.github.dakusui.pcond.functions.Functions.*;
-import static com.github.dakusui.pcond.functions.Predicates.*;
+import static com.github.dakusui.pcond.forms.Functions.*;
+import static com.github.dakusui.pcond.forms.Predicates.*;
 import static com.github.dakusui.pcond.ut.bugfixes.issue42.Issue42Utils.createPrivateExtendingPublicMap;
 import static org.junit.Assert.assertEquals;
